Added content_type and friend to Posts
This update allows you to store who sent the post and who (if anyone) is to
receive it. By do this and added a Friend visiblilty you can now send posts
that are private to one other author. This commit also includes a new
attribute for content_type, where you can specify markdown or plain text